+++ tests/WebIMTest.php
@@ -16,7 +16,7 @@ class WebIMTest extends PHPUnit_Framework_TestCase {
         $apikey   = 'public';
         $server   = 'http://localhost:8000';
         $buddies  = ['uid2', 'uid3'];
-        $groups   = ['gid1', 'gid2'];
+        $rooms    = ['room1', 'room2'];
         $endpoint = array(
             'id' => 'uid1',
             'nick' => 'User1',
@@ -24,11 +24,11 @@ class WebIMTest extends PHPUnit_Framework_TestCase {
             'status' => 'online',
         );
         $this->webim = new \WebIM\WebIM($endpoint, $domain, $apikey, $server);
-        $this->webim->online($buddies, $groups);
+        $this->webim->online($buddies, $rooms);
     }
 
     public function testOnline() {
-        $this->dump('Online', $this->webim->online(['uid4', 'uid5'], ['gid6', 'gid7']));
+        $this->dump('Online', $this->webim->online(['uid4', 'uid5'], ['room6', 'room7']));
     }
 
     public function testShow() {
@@ -56,16 +56,16 @@ class WebIMTest extends PHPUnit_Framework_TestCase {
     }
 
     public function testMembers() {
-        $this->dump('Members', $this->webim->members('gid1'));
+        $this->dump('Members', $this->webim->members('room1'));
     }
 
     public function testJoin() {
-        $this->dump('Join', $this->webim->join('gid3'));
+        $this->dump('Join', $this->webim->join('room3'));
     }
 
     public function testLeave() {
-        $this->webim->join('gid3');
-        $this->dump('Leave', $this->webim->leave('gid3'));
+        $this->webim->join('room3');
+        $this->dump('Leave', $this->webim->leave('room3'));
     }
 
     private function dump($title, $data) {
@@ -73,3 +73,4 @@ class WebIMTest extends PHPUnit_Framework_TestCase {
     }
 
 }
+